# onboarding/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Interactive onboarding platform: tours, immutable versions, resolve and analytics.",
        "title": "Cluer Onboarding",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/tours": {
            "get": {
                "description": "Returns every tour of an app with the status of its current version",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tours"
                ],
                "summary": "List tours",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "App ID",
                        "name": "appId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Tour"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a tour together with its first draft version",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tours"
                ],
                "summary": "Create a new tour",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "App ID",
                        "name": "appId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Tour object",
                        "name": "tour",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Tour"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Tour"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            }
        },
        "/tours/published": {
            "get": {
                "description": "Returns published tours with their hints for a target path",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tours"
                ],
                "summary": "Get published tours by path",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "App ID",
                        "name": "appId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "/dashboard",
                        "description": "Target path",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Tour"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            }
        },
        "/tours/{tourId}": {
            "get": {
                "description": "Returns the tour with both its published and draft versions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tours"
                ],
                "summary": "Get a tour card",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tour ID",
                        "name": "tourId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.TourCard"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Soft delete: the tour stops resolving but its analytics history survives",
                "tags": [
                    "Tours"
                ],
                "summary": "Archive a tour",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tour ID",
                        "name": "tourId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Changes title, description, enabled or priority without creating a version",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tours"
                ],
                "summary": "Update tour metadata",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tour ID",
                        "name": "tourId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to change",
                        "name": "patch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.TourMetaPatch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Tour"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/tours/{tourId}/draft": {
            "post": {
                "description": "Entry point to editing: the published version itself is never modified",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Versions"
                ],
                "summary": "Fork a draft from the published version",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tour ID",
                        "name": "tourId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.TourVersion"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Saves work in progress; validation is enforced at publish time, not here",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Versions"
                ],
                "summary": "Update the draft content",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tour ID",
                        "name": "tourId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to change",
                        "name": "patch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DraftPatch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.TourVersion"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/tours/{tourId}/publish": {
            "post": {
                "description": "Archives the live version and promotes the draft, atomically",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Versions"
                ],
                "summary": "Publish the draft",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tour ID",
                        "name": "tourId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.TourVersion"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            }
        },
        "/tours/{tourId}/rollback": {
            "post": {
                "description": "Copies the chosen version into a new one and publishes it; history is appended, never rewritten",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Versions"
                ],
                "summary": "Roll back to an earlier version",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tour ID",
                        "name": "tourId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Target version",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RollbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.TourVersion"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            }
        },
        "/tours/{tourId}/versions": {
            "get": {
                "description": "Version history without hint bodies, newest first",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Versions"
                ],
                "summary": "List tour versions",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tour ID",
                        "name": "tourId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.TourVersion"
                            }
                        }
                    }
                }
            }
        },
...
        "/versions/{versionId}": {
            "get": {
                "description": "Returns a version with its hints, for diffing and preview",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Versions"
                ],
                "summary": "Get one version",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Version ID",
                        "name": "versionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.TourVersion"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errs.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Shorthand error",
                    "type": "string"
                },
                "message": {
                    "description": "Pretty error message",
                    "type": "string"
                }
            }
        },
...
        "request.DraftPatch": {
            "type": "object",
            "properties": {
                "audience": {
                    "$ref": "#/definitions/request.Audience"
                },
                "target_path": {
                    "type": "string"
                },
                "trigger_config": {
                    "$ref": "#/definitions/request.TriggerConfig"
                },
                "trigger_type": {
                    "type": "string"
                }
            }
        },
...
        "request.Tour": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "audience": {
                    "$ref": "#/definitions/request.Audience"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "description": {
                    "type": "string",
                    "example": "Тур по основным функциям дашборда"
                },
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "hints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.Hint"
                    }
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "priority": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "draft",
                        "published",
                        "archived"
                    ],
                    "example": "draft"
                },
                "target_path": {
                    "type": "string",
                    "example": "/dashboard"
                },
                "title": {
                    "type": "string",
                    "example": "Добро пожаловать в систему"
                },
                "trigger_config": {
                    "$ref": "#/definitions/request.TriggerConfig"
                },
                "trigger_type": {
                    "type": "string",
                    "enum": [
                        "on_load",
                        "delay",
                        "exit_intent",
                        "manual",
                        "scroll_depth",
                        "inactivity",
                        "element_visible"
                    ],
                    "example": "on_load"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "version": {
                    "type": "integer",
                    "example": 7
                },
                "version_id": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        },
        "request.TourMetaPatch": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "priority": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "response.Tour": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "audience": {
                    "$ref": "#/definitions/response.Audience"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "description": {
                    "type": "string",
                    "example": "Тур по основным функциям дашборда"
                },
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "hints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Hint"
                    }
                },
                "id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "01914f6a-8c9b-7a0b-9b0b-8c9b7a0b9b0b"
                },
                "priority": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "draft",
                        "published",
                        "archived"
                    ],
                    "example": "draft"
                },
                "target_path": {
                    "type": "string",
                    "example": "/dashboard"
                },
                "title": {
                    "type": "string",
                    "example": "Добро пожаловать в систему"
                },
                "trigger_config": {
                    "$ref": "#/definitions/response.TriggerConfig"
                },
                "trigger_type": {
                    "type": "string",
                    "enum": [
                        "on_load",
                        "delay",
                        "exit_intent",
                        "manual",
                        "scroll_depth",
                        "inactivity",
                        "element_visible"
                    ],
                    "example": "on_load"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "version": {
                    "type": "integer",
                    "example": 7
                },
                "version_id": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        },
        "response.TourCard": {
            "type": "object",
            "properties": {
                "draft": {
                    "$ref": "#/definitions/response.TourVersion"
                },
                "published": {
                    "$ref": "#/definitions/response.TourVersion"
                },
                "tour": {
                    "$ref": "#/definitions/response.Tour"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "OIDC id token: \"Bearer \u003cid_token\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}